tried rake task and background job to no avail
