+++ test/jobs/scrape_job_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class ScrapeJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
